requirements.txt
langgraph
langchain
langchain-google-genai
langchain-ollama

fastapi
uvicorn
python-dotenv
httpx
cachetools
pydantic